# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ForensiX Digital Forensics Investigation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List all cases
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Case"
    post:
      operationId: createCase
      tags: [cases]
      summary: Create a new case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseInput"
      responses:
        "201":
          description: Case created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
  /cases/{caseId}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get a case by ID
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        "404":
          description: Case not found
    patch:
      operationId: updateCase
      tags: [cases]
      summary: Update a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseUpdate"
      responses:
        "200":
          description: Case updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        "404":
          description: Case not found
    delete:
      operationId: deleteCase
      tags: [cases]
      summary: Delete a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Case deleted
        "404":
          description: Case not found
  /cases/{caseId}/evidence:
    get:
      operationId: listCaseEvidence
      tags: [evidence]
      summary: List evidence for a case
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of evidence
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Evidence"
    post:
      operationId: uploadEvidence
      tags: [evidence]
      summary: Upload evidence file for a case (multipart/form-data with field 'file')
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Evidence uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidence"
        "404":
          description: Case not found
...
  /evidence/{evidenceId}:
    get:
      operationId: getEvidence
      tags: [evidence]
      summary: Get evidence by ID
      parameters:
        - name: evidenceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Evidence details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidence"
        "404":
          description: Evidence not found
    delete:
      operationId: deleteEvidence
      tags: [evidence]
      summary: Delete evidence
      parameters:
        - name: evidenceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Evidence deleted
        "404":
          description: Evidence not found
  /evidence/{evidenceId}/analyze:
    post:
      operationId: analyzeEvidence
      tags: [evidence]
      summary: Trigger analysis of an evidence file
      parameters:
        - name: evidenceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "404":
          description: Evidence not found
  /evidence/{evidenceId}/analysis:
    get:
      operationId: getEvidenceAnalysis
      tags: [evidence]
      summary: Get analysis results for evidence
      parameters:
        - name: evidenceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "404":
          description: Evidence not found
  /evidence/{evidenceId}/custody:
    get:
      operationId: getChainOfCustody
      tags: [evidence]
      summary: Get chain of custody for evidence
      parameters:
        - name: evidenceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chain of custody records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustodyRecord"
...
components:
  schemas:
...
    Case:
      type: object
      required: [id, caseNumber, name, status, investigator, createdAt, updatedAt, evidenceCount]
      properties:
        id:
          type: integer
        caseNumber:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, active, closed, archived]
        priority:
          type: ["string", "null"]
          enum: [low, medium, high, critical, null]
        investigator:
          type: string
        tags:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        evidenceCount:
          type: integer
    CaseInput:
      type: object
      required: [name, investigator]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        investigator:
          type: string
          minLength: 1
        priority:
          type: string
          enum: [low, medium, high, critical]
        tags:
          type: string
    CaseUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [open, active, closed, archived]
        investigator:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        tags:
          type: string
    Evidence:
      type: object
      required: [id, caseId, originalName, storedName, mimeType, fileSize, sha256, md5, sha1, analysisStatus, uploadedAt]
      properties:
        id:
          type: integer
        caseId:
          type: integer
        originalName:
          type: string
        storedName:
          type: string
        mimeType:
          type: string
        fileSize:
          type: integer
        sha256:
          type: string
        md5:
          type: string
        sha1:
          type: string
        analysisStatus:
          type: string
          enum: [pending, analyzing, complete, failed]
        notes:
          type: ["string", "null"]
        uploadedAt:
          type: string
    AnalysisResult:
      type: object
      required: [evidenceId, status]
      properties:
        evidenceId:
          type: integer
        status:
          type: string
        fileMetadata:
          $ref: "#/components/schemas/FileMetadata"
        strings:
          $ref: "#/components/schemas/ExtractedStrings"
        documentMeta:
          $ref: "#/components/schemas/DocumentMeta"
        imageMeta:
          $ref: "#/components/schemas/ImageMeta"
        emailMeta:
          $ref: "#/components/schemas/EmailMeta"
        findings:
          type: array
          items:
            $ref: "#/components/schemas/Finding"
    FileMetadata:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        mimeType:
          type: string
        created:
          type: ["string", "null"]
        modified:
          type: ["string", "null"]
        accessed:
          type: ["string", "null"]
        sha256:
          type: string
        md5:
          type: string
        sha1:
          type: string
    ExtractedStrings:
      type: object
      properties:
        urls:
          type: array
          items:
            type: string
        ipAddresses:
          type: array
          items:
            type: string
        emailAddresses:
          type: array
          items:
            type: string
        domains:
          type: array
          items:
            type: string
        printableStrings:
          type: array
          items:
            type: string
    DocumentMeta:
      type: object
      properties:
        author:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        creator:
          type: ["string", "null"]
        producer:
          type: ["string", "null"]
        creationDate:
          type: ["string", "null"]
        modDate:
          type: ["string", "null"]
        pageCount:
          type: ["integer", "null"]
        wordCount:
          type: ["integer", "null"]
    ImageMeta:
      type: object
      properties:
        make:
          type: ["string", "null"]
        model:
          type: ["string", "null"]
        software:
          type: ["string", "null"]
        dateTime:
          type: ["string", "null"]
        gpsLatitude:
          type: ["number", "null"]
        gpsLongitude:
          type: ["number", "null"]
        width:
          type: ["integer", "null"]
        height:
          type: ["integer", "null"]
        colorSpace:
          type: ["string", "null"]
    EmailMeta:
      type: object
      properties:
        from:
          type: ["string", "null"]
        to:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        date:
          type: ["string", "null"]
        messageId:
          type: ["string", "null"]
        receivedChain:
          type: array
          items:
            type: string
        hasAttachments:
          type: boolean
        attachmentNames:
          type: array
          items:
            type: string
    Finding:
      type: object
      required: [id, evidenceId, severity, description, indicatorType, value, createdAt]
      properties:
        id:
          type: integer
        evidenceId:
          type: integer
        caseId:
          type: integer
        severity:
          type: string
          enum: [info, low, medium, high, critical]
        description:
          type: string
        indicatorType:
          type: string
        value:
          type: string
        createdAt:
          type: string
...
    CustodyRecord:
      type: object
      required: [id, evidenceId, action, actor, timestamp]
      properties:
        id:
          type: integer
        evidenceId:
          type: integer
        action:
          type: string
        actor:
          type: string
        notes:
          type: ["string", "null"]
        timestamp:
          type: string
...