# lib/api-zod/src/generated/types/case.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ForensiX Digital Forensics Investigation Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CasePriority } from './casePriority';
import type { CaseStatus } from './caseStatus';

export interface Case {
  id: number;
  caseNumber: string;
  name: string;
  /** @nullable */
  description?: string | null;
  status: CaseStatus;
  /** @nullable */
  priority?: CasePriority;
  investigator: string;
  /** @nullable */
  tags?: string | null;
  createdAt: string;
  updatedAt: string;
  evidenceCount: number;
}