# lib/api-zod/src/generated/types/emailMeta.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ForensiX Digital Forensics Investigation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface EmailMeta {
  /** @nullable */
  from?: string | null;
  /** @nullable */
  to?: string | null;
  /** @nullable */
  subject?: string | null;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  messageId?: string | null;
  receivedChain?: string[];
  hasAttachments?: boolean;
  attachmentNames?: string[];
}